bug: Fixed the current display
