+++ src/const/entries.ts
@@ -1,0 +1,6 @@
+export const ENTRIES = [
+  { href: '/', content: 'home' },
+  { href: '/blog', content: 'blog' },
+  { href: '/about', content: 'about' },
+  { href: '/contact', content: 'contact' },
+];
